decoy‑honeypot/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
python-docx==1.1.2
typing_extensions==4.13.2
Werkzeug==3.1.3
